Real-Time Stock Market Dashboard/requirements.txt
streamlit==1.28.0
pandas==2.1.0
plotly==5.17.0
yfinance==0.2.28
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
ta==0.10.2  # Technical Analysis library